pyproject.toml
[project]
name = "meg-speech-decode"
version = "0.1.0"
description = "Baseline pipeline for speech decoding from MEG recordings"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mne>=1.6.0",
  "librosa>=0.10.0",
  "numpy>=1.25.0",
  "pandas>=2.0.0",
  "scikit-learn>=1.4.0",
  "scipy>=1.11.0",
  "PyYAML>=6.0",
  "tqdm>=4.66.0",
  "joblib>=1.3.0",
  "gdown>=5.2.0"
]

[project.optional-dependencies]
deep = [
  "torch>=2.2.0",
  "torchaudio>=2.2.0",
  "transformers>=4.40.0",
  "sentencepiece>=0.2.0"
]

[project.scripts]
meg-decode = "meg_speech_decode.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
